pyproject.toml
[project]
name = "lkeep"
version = "0.1.0"
description = ""
authors = [
    { name = "proDream", email = "sushkoos@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.115.6,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "pre-commit (>=4.0.1,<5.0.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "alembic (>=1.14.0,<2.0.0)",
    "ruff (>=0.9.0,<0.10.0)",
    "pydantic[email] (>=2.10.5,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (==4.0.1)",
    "celery (>=5.4.0,<6.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "lkeep", from = "src"}]

[project.scripts]
app = "lkeep.main:start"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
